refactor: extract check command into its own module (#1350)
Move the check command's orchestration logic, output analysis types,
and parsing functions from cli.rs into a dedicated check/ directory,
following the same pattern as exec/.

cli.rs: 1,778 → 1,314 lines
check/mod.rs: execute_check() orchestration (247 lines)
check/analysis.rs: types, parsers, formatters (255 lines)

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Medium Risk**
> Primarily a refactor, but it changes how `vp check` is dispatched and how fmt/lint output is captured/parsed, which could affect exit codes or CI output stability.
>
> **Overview**
> Refactors the `vp check` implementation by extracting its orchestration and output-parsing logic from `cli.rs` into a new `check/` module (`check/mod.rs` + `check/analysis.rs`), keeping the same fmt→lint flow and `--fix` re-format pass.
>
> Updates `cli.rs` visibility to reuse `resolve_universal_vite_config` and `resolve_and_capture_output`, and switches the `Check` subcommand handler to delegate to `crate::check::execute_check`; related `LintMessageKind` tests move with the extraction. Documentation in `rfcs/check-command.md` is updated to reflect the new module layout and cache-input helper naming.
>
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit e9cf38edebf5f5dc418b10ce3fc3f4e0497052ed. Configure [here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->
